internal/doc/generate: move to config/doc
We need to export it so Butane can use it.
